feat(core): split index.ts of monorepo-scanner
